backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-docx==1.1.2
PyPDF2==3.0.1
openpyxl==3.1.5
groq==0.11.0
numpy==1.26.4
scikit-learn==1.5.2
aiofiles==24.1.0
python-dotenv==1.0.1
reportlab==4.2.5
